pyproject.toml
[project]
name = "tasty-agent"
dynamic = ["version"]
description = "A Model Context Protocol server for interacting with TastyTrade brokerage accounts"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=v1.14.0",
    "tastytrade>=11.0.0",
    "humanize>=4.12.3",
    "aiolimiter==1.2.1",
    "tabulate>=0.9.0",
    "aiocache>=0.12.2",
    "python-dotenv>=1.1.0",
]
authors = [{ name = "Ferdous", email = "ferdous@hey.com" }]
keywords = ["tastytrade", "trading", "mcp", "model context protocol", "broker", "robinhood", "wallstreetbets"]
classifiers = [
    "Intended Audience :: Financial and Insurance Industry",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Office/Business :: Financial",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]

[project.scripts]
tasty-agent = "tasty_agent.server:mcp_app.run"
tasty-agent-http = "tasty_agent.http_server:main"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.targets.wheel]
packages = ["tasty_agent"]

[tool.hatch.build.hooks.vcs]
version-file = "tasty_agent/_version.py"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pydantic-ai>=0.4.6",
    "python-dotenv>=1.1.0",
    "typer",
]
